undo renaming
